implement CRUD operations in frontend
