+++ .angular/cache/20.3.2/to-do-list-angular/vite/deps/@angular_platform-browser.js
@@ -34,10 +34,12 @@ import {
   withI18nSupport,
   withIncrementalHydration,
   withNoHttpTransferCache
-} from "./chunk-6SUY2G2I.js";
+} from "./chunk-B4ONJZK3.js";
+import "./chunk-KKPQKFER.js";
 import {
   getDOM
-} from "./chunk-FLER7ODA.js";
+} from "./chunk-SESEHIUB.js";
+import "./chunk-ZJ25XCV3.js";
 import "./chunk-NLZIVNDD.js";
 import "./chunk-WDMUDEB6.js";
 export {
