+++ .angular/cache/20.3.2/to-do-list-angular/vite/deps/@angular_router.js
@@ -1,6 +1,7 @@
 import {
   Title
-} from "./chunk-6SUY2G2I.js";
+} from "./chunk-B4ONJZK3.js";
+import "./chunk-KKPQKFER.js";
 import {
   HashLocationStrategy,
   LOCATION_INITIALIZED,
@@ -8,7 +9,8 @@ import {
   LocationStrategy,
   PathLocationStrategy,
   ViewportScroller
-} from "./chunk-FLER7ODA.js";
+} from "./chunk-SESEHIUB.js";
+import "./chunk-ZJ25XCV3.js";
 import {
   APP_BOOTSTRAP_LISTENER,
   ApplicationRef,
