+++ .angular/cache/20.3.2/to-do-list-angular/vite/deps/chunk-ZJ25XCV3.js
@@ -1,0 +1,29 @@
+// node_modules/@angular/common/fesm2022/xhr.mjs
+function parseCookieValue(cookieStr, name) {
+  name = encodeURIComponent(name);
+  for (const cookie of cookieStr.split(";")) {
+    const eqIndex = cookie.indexOf("=");
+    const [cookieName, cookieValue] = eqIndex == -1 ? [cookie, ""] : [cookie.slice(0, eqIndex), cookie.slice(eqIndex + 1)];
+    if (cookieName.trim() === name) {
+      return decodeURIComponent(cookieValue);
+    }
+  }
+  return null;
+}
+var XhrFactory = class {
+};
+
+export {
+  parseCookieValue,
+  XhrFactory
+};
+/*! Bundled license information:
+
+@angular/common/fesm2022/xhr.mjs:
+  (**
+   * @license Angular v20.3.1
+   * (c) 2010-2025 Google LLC. https://angular.io/
+   * License: MIT
+   *)
+*/
+//# sourceMappingURL=chunk-ZJ25XCV3.js.map
