+++ src/environments/environment.prod.ts
@@ -1,0 +1,5 @@
+export const environment = {
+  production: true,
+  apiUrl: "https://yourserver.com/tasks",
+};
+// Replace with your production API URL
