+++ src/environments/environments.ts
@@ -1,0 +1,4 @@
+export const environment = {
+  production: false,
+  apiUrl: "http://localhost:3000/tasks",
+};
